Fix/#3246 reduced the number of generated events 2

+++ src/com/serotonin/mango/rt/dataSource/modbus/ModbusIpDataSource.java
@@ -68,7 +68,7 @@ public class ModbusIpDataSource extends ModbusDataSource {
 
 				List<DataPointVO> points;
 				if (DataSourcePointsCache.getInstance().isCacheEnabled()) {
-					points = DataSourcePointsCache.getInstance().getDataPoints((long) configuration.getId());
+					points = DataSourcePointsCache.getInstance().getDataPoints(configuration.getId());
 				} else {
 					points = dataPointDao.getDataPoints(
 						configuration.getId(), null);
